Enhance type annotations and memory profiling output
- Added optional type annotations to `Statement::Let` for better type checking.
- Improved error handling for type mismatches during variable declaration.
- Enhanced memory profiling output for clarity and conciseness.
- Updated CLI messages for profiling to provide clearer context and formatting.
