+++ src/codegen/cranelift/lowering.rs
@@ -1,549 +1,29 @@
 use std::collections::HashMap;
-use std::fs;
-use std::path::Path;
-use std::process::Command;
+use std::hash::{Hash, Hasher};
 
 use anyhow::{Context, Result, anyhow, bail};
-use cranelift_codegen::ir::InstBuilder;
-use cranelift_codegen::{Context, ir, isa, settings};
-use cranelift_frontend::{FunctionBuilder, FunctionBuilderContext};
-use cranelift_jit::{JITBuilder, JITModule};
-use cranelift_module::{DataId, FuncId, Linkage, Module};
-use cranelift_object::{ObjectBuilder, ObjectModule};
-use target_lexicon::Triple;
-use tracing::warn;
+use cranelift_codegen::{ir, ir::InstBuilder};
+use cranelift_frontend::FunctionBuilder;
 
-use crate::codegen::llvm::{preferred_target_flag, prepare_rust_bridges};
-use crate::codegen::target::TargetTriple;
-use crate::codegen::{BuildArtifact, CodegenOptions};
-use crate::runtime::symbol_registry::SymbolRegistry;
-use crate::typecheck::TypeInfo;
-use ast::nodes::{
-    BinaryOp, Expr, FStringPart, Function, Literal, Program, Statement, Type, UnaryOp,
-};
+use ast::nodes::{BinaryOp, Block, Expr, FStringPart, Literal, Pattern, Statement, UnaryOp};
 
-/// Build a native executable using the Cranelift backend
-pub fn build_executable(
-    program: &Program,
-    expr_types: &HashMap<usize, TypeInfo>,
-    output: &Path,
-    options: &CodegenOptions,
-) -> Result<BuildArtifact> {
-    let registry = crate::runtime::ffi::bootstrap_stdlib();
-    let bridge_libraries = prepare_rust_bridges(program, registry)?;
-
-    if options.emit_ir {
-        warn!(
-            "Cranelift backend does not currently support emitting textual IR; ignoring --dump-ir"
-        );
-    }
-    if options.enable_lto {
-        warn!("Cranelift backend does not support LTO; disabling -flto for this build");
-    }
-    if options.enable_pgo {
-        warn!("Cranelift backend does not support PGO; ignoring PGO-related flags");
-    }
-
-    // Ensure we have an entry point for executable builds
-    if !program.functions().any(|f| f.name == "main") {
-        bail!("entry function `main` not found");
-    }
-
-    let runtime_triple = resolve_runtime_triple(options);
-    let triple_str = runtime_triple.to_llvm_triple();
-
-    let mut backend =
-        CraneliftBackend::new(&runtime_triple, registry, clone_expr_type_map(expr_types))?;
-
-    let mut declared = Vec::new();
-    for function in program.functions() {
-        let func_id = backend.declare_function(function)?;
-        declared.push((func_id, function));
-    }
-    for (func_id, function) in declared {
-        backend.build_function(func_id, function)?;
-    }
-
-    let compiled = backend.finalize_module()?;
-    let object_bytes = compiled
-        .object_bytes
-        .ok_or_else(|| anyhow!("Cranelift backend did not produce object bytes"))?;
-
-    let object_path = output.with_extension("o");
-    fs::write(&object_path, &object_bytes).with_context(|| {
-        format!(
-            "failed to write Cranelift object file {}",
-            object_path.display()
-        )
-    })?;
-
-    let runtime_c = if runtime_triple.is_wasm() {
-        None
-    } else {
-        let runtime_c = output.with_extension("runtime.c");
-        let runtime_c_content = runtime_triple.runtime_c_code();
-        fs::write(&runtime_c, runtime_c_content)
-            .context("failed to write runtime C shim for Cranelift backend")?;
-        Some(runtime_c)
-    };
-
-    let runtime_o = if let Some(ref rt_c) = runtime_c {
-        let runtime_o = output.with_extension("runtime.o");
-        let c_compiler = runtime_triple.c_compiler();
-        let mut cc = Command::new(&c_compiler);
-        cc.arg("-c");
-        if runtime_triple.needs_pic() && !runtime_triple.is_windows() {
-            cc.arg("-fPIC");
-        }
-        let compiler_target_flag = preferred_target_flag(&c_compiler);
-        cc.arg(compiler_target_flag).arg(&triple_str);
-        cc.arg(rt_c).arg("-o").arg(&runtime_o);
-
-        let cc_status = cc
-            .status()
-            .context("failed to compile Cranelift runtime C shim")?;
-        if !cc_status.success() {
-            bail!("failed to compile runtime C shim for Cranelift backend");
-        }
-        Some(runtime_o)
-    } else {
-        None
-    };
-
-    let linker = runtime_triple.linker();
-    let mut cc = Command::new(&linker);
-    let linker_target_flag = preferred_target_flag(&linker);
-
-    if runtime_triple.is_wasm() {
-        cc.arg(linker_target_flag)
-            .arg(&triple_str)
-            .arg("--no-entry")
-            .arg("--export-dynamic")
-            .arg(&object_path)
-            .arg("-o")
-            .arg(output);
-    } else {
-        cc.arg(linker_target_flag).arg(&triple_str);
-        cc.arg(&object_path);
-        if let Some(ref rt_o) = runtime_o {
-            cc.arg(rt_o);
-        }
-        cc.arg("-o").arg(output);
-    }
-
-    for flag in runtime_triple.linker_flags() {
-        cc.arg(&flag);
-    }
-
-    for lib in &bridge_libraries {
-        cc.arg(lib);
-    }
-
-    let status = cc.status().context("failed to link Cranelift executable")?;
-
-    if !status.success() {
-        bail!("linker invocation failed with status {status}");
-    }
-
-    if let Some(ref rt_c) = runtime_c {
-        fs::remove_file(rt_c).ok();
-    }
-    if let Some(ref rt_o) = runtime_o {
-        fs::remove_file(rt_o).ok();
-    }
-    fs::remove_file(&object_path).ok();
-
-    Ok(BuildArtifact {
-        binary: output.to_path_buf(),
-        ir: None,
-    })
-}
-
-/// Build a shared library using the Cranelift backend
-pub fn build_shared_library(
-    program: &Program,
-    expr_types: &HashMap<usize, TypeInfo>,
-    output: &Path,
-    options: &CodegenOptions,
-) -> Result<BuildArtifact> {
-    let registry = crate::runtime::ffi::bootstrap_stdlib();
-    let bridge_libraries = prepare_rust_bridges(program, registry)?;
-
-    if options.emit_ir {
-        warn!(
-            "Cranelift backend does not currently support emitting textual IR; ignoring --dump-ir"
-        );
-    }
-    if options.enable_lto {
-        warn!("Cranelift backend does not support LTO; disabling -flto for this build");
-    }
-    if options.enable_pgo {
-        warn!("Cranelift backend does not support PGO; ignoring PGO-related flags");
-    }
-
-    let runtime_triple = resolve_runtime_triple(options);
-    let triple_str = runtime_triple.to_llvm_triple();
-
-    let mut backend =
-        CraneliftBackend::new(&runtime_triple, registry, clone_expr_type_map(expr_types))?;
-
-    // Two-pass lowering: declare all functions before defining bodies
-    let mut declared = Vec::new();
-    for function in program.functions() {
-        let func_id = backend.declare_function(function)?;
-        declared.push((func_id, function));
-    }
-    for (func_id, function) in declared {
-        backend.build_function(func_id, function)?;
-    }
-
-    let compiled = backend.finalize_module()?;
-    let object_bytes = compiled
-        .object_bytes
-        .ok_or_else(|| anyhow!("Cranelift backend did not produce object bytes"))?;
-
-    let object_path = output.with_extension("o");
-    fs::write(&object_path, &object_bytes).with_context(|| {
-        format!(
-            "failed to write Cranelift object file {}",
-            object_path.display()
-        )
-    })?;
-
-    let runtime_c = if runtime_triple.is_wasm() {
-        None
-    } else {
-        let runtime_c = output.with_extension("runtime.c");
-        let runtime_c_content = runtime_triple.runtime_c_code();
-        fs::write(&runtime_c, runtime_c_content)
-            .context("failed to write runtime C shim for Cranelift backend")?;
-        Some(runtime_c)
-    };
-
-    let runtime_o = if let Some(ref rt_c) = runtime_c {
-        let runtime_o = output.with_extension("runtime.o");
-        let c_compiler = runtime_triple.c_compiler();
-        let mut cc = Command::new(&c_compiler);
-        cc.arg("-c");
-        if runtime_triple.needs_pic() && !runtime_triple.is_windows() {
-            cc.arg("-fPIC");
-        }
-        let compiler_target_flag = preferred_target_flag(&c_compiler);
-        cc.arg(compiler_target_flag).arg(&triple_str);
-        cc.arg(rt_c).arg("-o").arg(&runtime_o);
-
-        let cc_status = cc
-            .status()
-            .context("failed to compile Cranelift runtime C shim")?;
-        if !cc_status.success() {
-            bail!("failed to compile runtime C shim for Cranelift backend");
-        }
-        Some(runtime_o)
-    } else {
-        None
-    };
-
-    let lib_ext = if runtime_triple.is_wasm() {
-        "wasm"
-    } else if runtime_triple.is_windows() {
-        "dll"
-    } else if runtime_triple.os == "darwin" {
-        "dylib"
-    } else {
-        "so"
-    };
-    let lib_path = if output
-        .extension()
-        .map(|ext| ext == lib_ext)
-        .unwrap_or(false)
-    {
-        output.to_path_buf()
-    } else {
-        output.with_extension(lib_ext)
-    };
-
-    let linker = runtime_triple.linker();
-    let mut cc = Command::new(&linker);
-    let linker_target_flag = preferred_target_flag(&linker);
-
-    if runtime_triple.is_wasm() {
-        cc.arg(linker_target_flag)
-            .arg(&triple_str)
-            .arg("--no-entry")
-            .arg("--export-dynamic")
-            .arg("-o")
-            .arg(&lib_path)
-            .arg(&object_path);
-    } else {
-        cc.arg("-shared");
-        if runtime_triple.needs_pic() {
-            cc.arg("-fPIC");
-        }
-        cc.arg(linker_target_flag).arg(&triple_str);
-        cc.arg("-o").arg(&lib_path).arg(&object_path);
-        if let Some(ref rt_o) = runtime_o {
-            cc.arg(rt_o);
-        }
-    }
-
-    for flag in runtime_triple.linker_flags() {
-        cc.arg(&flag);
-    }
-
-    for lib in &bridge_libraries {
-        cc.arg(lib);
-    }
-
-    let status = cc
-        .status()
-        .context("failed to invoke system linker for Cranelift backend")?;
-    if !status.success() {
-        bail!("Cranelift linker invocation failed with status {status}");
-    }
-
-    if let Some(ref rt_c) = runtime_c {
-        fs::remove_file(rt_c).ok();
-    }
-    if let Some(ref rt_o) = runtime_o {
-        fs::remove_file(rt_o).ok();
-    }
-    fs::remove_file(&object_path).ok();
-
-    Ok(BuildArtifact {
-        binary: lib_path,
-        ir: None,
-    })
-}
-
-fn resolve_runtime_triple(options: &CodegenOptions) -> TargetTriple {
-    if let Some(ref target) = options.target {
-        return target.clone();
-    }
-
-    let native = target_lexicon::Triple::host();
-    let native_str = native.to_string();
-    TargetTriple::parse(&native_str)
-        .unwrap_or_else(|_| TargetTriple::new("x86_64", "unknown", "linux", Some("gnu")))
-}
-
-fn clone_expr_type_map(expr_types: &HashMap<usize, TypeInfo>) -> HashMap<usize, TypeInfo> {
-    expr_types
-        .iter()
-        .map(|(id, ty)| (*id, ty.clone()))
-        .collect::<HashMap<_, _>>()
-}
-
-/// Backend-agnostic codegen trait
-pub trait CodegenBackend {
-    /// Declare a function with the given signature
-    fn declare_function(&mut self, function: &Function) -> Result<FuncId>;
-
-    /// Build the body of a declared function
-    fn build_function(&mut self, func_id: FuncId, function: &Function) -> Result<()>;
-
-    /// Finalize the module and return compiled artifacts
-    fn finalize_module(&mut self) -> Result<CompiledModule>;
-
-    /// Get information about the target platform
-    fn get_target_info(&self) -> TargetInfo;
-
-    /// Emit an object file to the given path
-    fn emit_object_file(&self, path: &Path) -> Result<()>;
-
-    /// Create a JIT context for runtime execution
-    fn create_jit_context(&self) -> Result<JitContext>;
-}
-
-/// Information about the compilation target
-#[derive(Debug, Clone)]
-pub struct TargetInfo {
-    pub triple: Triple,
-    pub pointer_width: u32,
-    pub endianness: target_lexicon::Endianness,
-}
-
-/// Compiled module containing functions and data
-#[derive(Debug)]
-pub struct CompiledModule {
-    pub functions: HashMap<String, FuncId>,
-    pub data_objects: HashMap<String, DataId>,
-    pub object_bytes: Option<Vec<u8>>,
-}
-
-/// JIT context for runtime execution
-#[derive(Debug)]
-pub struct JitContext {
-    /// The JIT module for dynamic code execution
-    pub module: JITModule,
-    /// Cache of declared functions in the JIT module
-    pub functions: HashMap<String, FuncId>,
-}
-
-/// Cranelift-based codegen backend
-pub struct CraneliftBackend {
-    /// The Cranelift module for managing functions and data
-    module: ObjectModule,
-    /// Function builder context for reusing allocations
-    fn_builder_ctx: FunctionBuilderContext,
-    /// Target ISA for code generation
-    isa: std::sync::Arc<dyn isa::TargetIsa>,
-    /// Symbol registry for FFI functions
-    _symbol_registry: &'static SymbolRegistry,
-    /// Type information from type checking
-    _expr_types: HashMap<usize, TypeInfo>,
-    /// Cache of declared functions
-    declared_functions: HashMap<String, FuncId>,
-    /// Cache of declared data objects
-    declared_data: HashMap<String, DataId>,
-}
+use super::backend::CraneliftBackend;
 
 impl CraneliftBackend {
-    /// Create a new Cranelift backend with the given configuration
-    pub fn new(
-        target: &TargetTriple,
-        symbol_registry: &'static SymbolRegistry,
-        expr_types: HashMap<usize, TypeInfo>,
-    ) -> Result<Self> {
-        // Convert our TargetTriple to target-lexicon Triple
-        let triple = Self::target_triple_to_lexicon(target)?;
-
-        // Create ISA builder for the target
-        let isa = isa::lookup(triple.clone())
-            .map_err(|e| anyhow!("Failed to create ISA for target {}: {}", triple, e))?
-            .finish(settings::Flags::new(settings::builder()))
-            .map_err(|e| anyhow!("Failed to finish ISA setup: {}", e))?;
-
-        // Create object module
-        let module = ObjectModule::new(
-            ObjectBuilder::new(
-                isa.clone(),
-                "otterlang".to_string(),
-                cranelift_module::default_libcall_names(),
-            )
-            .map_err(|e| anyhow!("Failed to create object builder: {}", e))?,
-        );
-
-        Ok(Self {
-            module,
-            fn_builder_ctx: FunctionBuilderContext::new(),
-            isa,
-            _symbol_registry: symbol_registry,
-            _expr_types: expr_types,
-            declared_functions: HashMap::new(),
-            declared_data: HashMap::new(),
-        })
-    }
-
-    /// Convert our TargetTriple to target-lexicon Triple
-    fn target_triple_to_lexicon(target: &TargetTriple) -> Result<Triple> {
-        let triple_str = format!("{}-{}-{}", target.arch, target.vendor, target.os);
-        triple_str
-            .parse()
-            .map_err(|e| anyhow!("Invalid target triple {}: {}", triple_str, e))
-    }
-
-    /// Convert Otter TypeInfo to Cranelift type
-    fn type_info_to_cranelift(&self, type_info: &TypeInfo) -> Result<ir::Type> {
-        match type_info {
-            TypeInfo::Unit => Ok(ir::types::INVALID), // Void/unit type
-            TypeInfo::Bool => Ok(ir::types::I8),
-            TypeInfo::I32 => Ok(ir::types::I32),
-            TypeInfo::I64 => Ok(ir::types::I64),
-            TypeInfo::F64 => Ok(ir::types::F64),
-            TypeInfo::Str => Ok(self.module.target_config().pointer_type()), // String as pointer
-            TypeInfo::List(_) | TypeInfo::Dict { .. } => {
-                // Collections are represented as pointers for now
-                Ok(self.module.target_config().pointer_type())
-            }
-            TypeInfo::Function { .. } => {
-                // Function pointers
-                Ok(self.module.target_config().pointer_type())
-            }
-            TypeInfo::Struct { .. } | TypeInfo::Enum { .. } => {
-                // Custom types are represented as pointers
-                // In a full implementation, we might want to handle structs differently
-                Ok(self.module.target_config().pointer_type())
-            }
-            TypeInfo::Generic { base, args } => {
-                // Handle generic types based on base type
-                match base.as_str() {
-                    "List" | "list" => Ok(self.module.target_config().pointer_type()),
-                    "Dict" | "dict" | "Map" | "map" => {
-                        Ok(self.module.target_config().pointer_type())
-                    }
-                    _ => {
-                        // For custom generic types, use pointer representation
-                        Ok(self.module.target_config().pointer_type())
-                    }
-                }
-            }
-            TypeInfo::Alias { underlying, .. } => {
-                // Follow the alias to the underlying type
-                self.type_info_to_cranelift(underlying)
-            }
-            TypeInfo::Unknown => {
-                // Unknown types default to pointer size
-                Ok(self.module.target_config().pointer_type())
-            }
-            TypeInfo::Error => {
-                // Error types are represented as strings
-                Ok(self.module.target_config().pointer_type())
-            }
-            TypeInfo::Module(_) => {
-                // Module types are opaque pointers
-                Ok(self.module.target_config().pointer_type())
-            }
-        }
-    }
-
-    /// Convert Otter AST Type to Cranelift type (legacy method for backward compatibility)
-    fn otter_type_to_cranelift(&self, otter_type: &Type) -> Result<ir::Type> {
-        // Convert AST Type to TypeInfo first, then to Cranelift type
-        let type_info = TypeInfo::from(otter_type);
-        self.type_info_to_cranelift(&type_info)
-    }
-
-    /// Build a function signature from Otter function
-    fn build_signature(&self, function: &Function) -> Result<ir::Signature> {
-        let mut sig = ir::Signature::new(self.isa.default_call_conv());
-
-        // Add parameters
-        for param in &function.params {
-            if let Some(ty) = &param.ty {
-                let param_type = self.otter_type_to_cranelift(ty)?;
-                sig.params.push(ir::AbiParam::new(param_type));
-            } else {
-                // Default to i64 if no type specified
-                sig.params.push(ir::AbiParam::new(ir::types::I64));
-            }
-        }
-
-        // Add return type
-        let ret_type = if let Some(ret_ty) = &function.ret_ty {
-            self.otter_type_to_cranelift(ret_ty)?
-        } else {
-            ir::types::I64 // Default return type
-        };
-
-        if ret_type != ir::types::INVALID {
-            sig.returns.push(ir::AbiParam::new(ret_type));
-        }
-
-        Ok(sig)
-    }
-
     /// Lower a block of statements to Cranelift IR
     fn lower_block_with_builder(
         &mut self,
         builder: &mut FunctionBuilder,
-        block: &ast::nodes::Block,
+        block: &Block,
         variables: &mut HashMap<String, ir::Value>,
     ) -> Result<ir::Value> {
         let mut last_value = None;
 
         for stmt in &block.statements {
             match stmt {
-                Statement::Let { name, expr, .. } => {
+                Statement::Let {
+                    name, ty: _, expr, ..
+                } => {
                     let val = self.lower_expr_with_builder(builder, expr, variables)?;
                     variables.insert(name.clone(), val);
                     last_value = Some(val);
@@ -2237,137 +1717,5 @@ impl CraneliftBackend {
         let result = builder.block_params(merge_block)[0];
 
         Ok(result)
-    }
-}
-
-impl CodegenBackend for CraneliftBackend {
-    fn declare_function(&mut self, function: &Function) -> Result<FuncId> {
-        let sig = self.build_signature(function)?;
-        let func_id = self
-            .module
-            .declare_function(&function.name, Linkage::Export, &sig)
-            .map_err(|e| anyhow!("Failed to declare function {}: {}", function.name, e))?;
-
-        self.declared_functions
-            .insert(function.name.clone(), func_id);
-        Ok(func_id)
-    }
-
-    fn build_function(&mut self, func_id: FuncId, function: &Function) -> Result<()> {
-        // Create a new function context
-        let mut ctx = Context::new();
-        ctx.func.signature = self.build_signature(function)?;
-
-        // Take the function builder context temporarily
-        let mut fn_builder_ctx = std::mem::take(&mut self.fn_builder_ctx);
-
-        // Create the function builder
-        let mut builder = FunctionBuilder::new(&mut ctx.func, &mut fn_builder_ctx);
-
-        // Create entry block
-        let entry_block = builder.create_block();
-        builder.switch_to_block(entry_block);
-
-        // Set up variables for function parameters
-        let mut variables = HashMap::new();
-
-        // Declare function parameters as variables
-        for (i, param) in function.params.iter().enumerate() {
-            let param_value = builder.block_params(entry_block)[i];
-            variables.insert(param.name.clone(), param_value);
-        }
-
-        // Seal the entry block before lowering (required by Cranelift)
-        builder.seal_block(entry_block);
-
-        // Lower the function body
-        let result =
-            Self::lower_block_with_builder(self, &mut builder, &function.body, &mut variables)?;
-
-        // Return the result of the last expression in the block
-        builder.ins().return_(&[result]);
-
-        // Finalize the function
-        builder.finalize();
-
-        // Restore the function builder context
-        self.fn_builder_ctx = fn_builder_ctx;
-
-        // Define the function in the module
-        self.module.define_function(func_id, &mut ctx)?;
-
-        Ok(())
-    }
-
-    fn finalize_module(&mut self) -> Result<CompiledModule> {
-        // Finalize the ObjectModule to get the object bytes
-        let product = self.module.finish();
-        let object_bytes = product
-            .object
-            .write()
-            .map_err(|e| anyhow!("Failed to write object: {}", e))?;
-
-        Ok(CompiledModule {
-            functions: self.declared_functions.clone(),
-            data_objects: self.declared_data.clone(),
-            object_bytes: Some(object_bytes),
-        })
-    }
-
-    fn get_target_info(&self) -> TargetInfo {
-        let triple = self.isa.triple().clone();
-        let pointer_width = self.isa.pointer_bits() as u32;
-        let endianness = match self.isa.endianness() {
-            ir::Endianness::Little => target_lexicon::Endianness::Little,
-            ir::Endianness::Big => target_lexicon::Endianness::Big,
-        };
-
-        TargetInfo {
-            triple,
-            pointer_width,
-            endianness,
-        }
-    }
-
-    fn emit_object_file(&self, path: &Path) -> Result<()> {
-        // First finalize the module to get object bytes
-        let product = self.module.finish();
-        let object_bytes = product
-            .object
-            .write()
-            .map_err(|e| anyhow!("Failed to write object: {}", e))?;
-
-        // Write the object bytes to the file
-        std::fs::write(path, object_bytes)
-            .map_err(|e| anyhow!("Failed to write object file to {}: {}", path.display(), e))?;
-
-        Ok(())
-    }
-
-    fn create_jit_context(&self) -> Result<JitContext> {
-        // Create JIT builder for the target
-        let jit_builder = JITBuilder::new(cranelift_module::default_libcall_names())
-            .map_err(|e| anyhow!("Failed to create JIT builder: {}", e))?;
-
-        // Create JIT module
-        let jit_module = JITModule::new(jit_builder);
-
-        // Copy over declared functions from the object module
-        let mut functions = HashMap::new();
-        for (name, &func_id) in &self.declared_functions {
-            let decl = self.module.declarations().get_function(func_id);
-
-            // Re-declare the function in the JIT module using the same linkage/signature
-            let jit_func_id = jit_module
-                .declare_function(name, decl.linkage, &decl.signature)
-                .map_err(|e| anyhow!("Failed to declare JIT function {}: {}", name, e))?;
-
-            functions.insert(name.clone(), jit_func_id);
-        }
-
-        Ok(JitContext {
-            module: jit_module,
-            functions,
-        })
     }
 }
