+++ src/codegen/cranelift/mod.rs
@@ -1,0 +1,7 @@
+pub mod backend;
+pub mod build;
+mod lowering;
+mod types;
+
+pub use backend::{CodegenBackend, CompiledModule, CraneliftBackend, JitContext, TargetInfo};
+pub use build::{build_executable, build_shared_library};
