+++ src/tools/profiler.rs
@@ -71,16 +71,15 @@ pub fn run_profiler() -> Result<()> {
 
 fn profile_memory(program: &Path, format: &str) -> Result<()> {
     println!(
-        "{} Starting memory profiling for: {}",
-        "🧠".cyan(),
-        program.display()
+        "{}",
+        format!("Starting memory profiling for: {}", program.display()).cyan()
     );
 
     let profiler = get_profiler();
     profiler.start();
 
     // Run the program (simplified - in practice would execute via runtime)
-    println!("{} Executing program...", "⚡".yellow());
+    println!("{}", "Executing program...".yellow());
     std::thread::sleep(Duration::from_millis(100)); // Simulate execution
 
     profiler.stop();
@@ -92,55 +91,25 @@ fn profile_memory(program: &Path, format: &str) -> Result<()> {
             println!("{}", json);
         }
         _ => {
-            println!("\n{} Memory Profiling Results:", "📊".magenta());
-            println!(
-                "  {} Total Allocated: {} bytes",
-                "💾".blue(),
-                stats.total_allocated
-            );
-            println!(
-                "  {} Total Freed: {} bytes",
-                "🗑️".green(),
-                stats.total_freed
-            );
-            println!(
-                "  {} Current Memory: {} bytes",
-                "📌".cyan(),
-                stats.current_memory
-            );
-            println!(
-                "  {} Peak Memory: {} bytes",
-                "📈".yellow(),
-                stats.peak_memory
-            );
-            println!(
-                "  {} Active Allocations: {}",
-                "🔢".white(),
-                stats.active_allocations
-            );
-            println!(
-                "  {} Duration: {:.2}s",
-                "⏱️".magenta(),
-                stats.duration_seconds
-            );
+            println!("\n{}", "Memory Profiling Results:".magenta());
+            println!("  Total Allocated: {} bytes", stats.total_allocated);
+            println!("  Total Freed: {} bytes", stats.total_freed);
+            println!("  Current Memory: {} bytes", stats.current_memory);
+            println!("  Peak Memory: {} bytes", stats.peak_memory);
+            println!("  Active Allocations: {}", stats.active_allocations);
+            println!("  Duration: {:.2}s", stats.duration_seconds);
 
             let leaks = profiler.detect_leaks();
             if !leaks.is_empty() {
                 println!(
-                    "\n{} Detected {} potential memory leaks:",
-                    "⚠️".red(),
+                    "\nDetected {} potential memory leaks:",
                     leaks.len()
                 );
                 for leak in leaks.iter().take(10) {
-                    println!(
-                        "  {} Address: 0x{:x}, Size: {} bytes",
-                        "🔍".yellow(),
-                        leak.pointer,
-                        leak.size
-                    );
+                    println!("  Address: 0x{:x}, Size: {} bytes", leak.pointer, leak.size);
                 }
             } else {
-                println!("\n{} No memory leaks detected!", "✅".green());
+                println!("\nNo memory leaks detected.");
             }
         }
     }
@@ -150,24 +119,20 @@ fn profile_memory(program: &Path, format: &str) -> Result<()> {
 
 fn profile_calls(program: &Path, iterations: usize) -> Result<()> {
     println!(
-        "{} Profiling function calls for: {}",
-        "🔬".cyan(),
-        program.display()
+        "{}",
+        format!("Profiling function calls for: {}", program.display()).cyan()
     );
-    println!("{} Iterations: {}", "🔄".yellow(), iterations);
+    println!("Iterations: {}", iterations);
 
     // In practice, would use CallProfiler
     // For now, provide a placeholder implementation
-    println!("\n{} Function Call Statistics:", "📊".magenta());
+    println!("\n{}", "Function Call Statistics:".magenta());
     println!(
         "{:<30} {:>12} {:>15} {:>15} {:>15}",
         "Function", "Calls", "Total Time", "Avg Time", "Max Time"
     );
     println!("{}", "-".repeat(90));
-    println!(
-        "{} Full call profiling requires runtime instrumentation",
-        "💡".yellow()
-    );
+    println!("{}", "Full call profiling requires runtime instrumentation".yellow());
 
     Ok(())
 }
@@ -178,31 +143,15 @@ fn show_stats(file: Option<PathBuf>) -> Result<()> {
             .with_context(|| format!("Failed to read stats file: {}", path.display()))?;
         let stats: ProfilingStats = serde_json::from_str(&content)?;
 
-        println!("{} Profiling Statistics:", "📊".magenta());
-        println!(
-            "  {} Duration: {:.2}s",
-            "⏱️".yellow(),
-            stats.duration_seconds
-        );
-        println!(
-            "  {} Total Allocated: {} bytes",
-            "💾".blue(),
-            stats.total_allocated
-        );
-        println!(
-            "  {} Total Freed: {} bytes",
-            "🗑️".green(),
-            stats.total_freed
-        );
-        println!(
-            "  {} Peak Memory: {} bytes",
-            "📈".yellow(),
-            stats.peak_memory
-        );
+        println!("{}", "Profiling Statistics:".magenta());
+        println!("  Duration: {:.2}s", stats.duration_seconds);
+        println!("  Total Allocated: {} bytes", stats.total_allocated);
+        println!("  Total Freed: {} bytes", stats.total_freed);
+        println!("  Peak Memory: {} bytes", stats.peak_memory);
     } else {
         println!(
-            "{} No stats file provided. Use --file to specify a stats file.",
-            "💡".yellow()
+            "{}",
+            "No stats file provided. Use --file to specify a stats file.".yellow()
         );
     }
     Ok(())
